Rename classes and exe
