+++ include/ShieldModelDetectorConstruction.hh
@@ -24,46 +24,35 @@
 // ********************************************************************
 //
 //
-/// \file B1PrimaryGeneratorAction.hh
-/// \brief Definition of the B1PrimaryGeneratorAction class
+/// \file ShieldModelDetectorConstruction.hh
+/// \brief Definition of the ShieldModelDetectorConstruction class
 
-#ifndef B1PrimaryGeneratorAction_h
-#define B1PrimaryGeneratorAction_h 1
+#ifndef ShieldModelDetectorConstruction_h
+#define ShieldModelDetectorConstruction_h 1
 
-#include "G4VUserPrimaryGeneratorAction.hh"
-#include "G4ParticleGun.hh"
-#include "G4GeneralParticleSource.hh"
+#include "G4VUserDetectorConstruction.hh"
 #include "globals.hh"
 
-class G4ParticleGun;
-class G4GeneralParticleSource;
-class G4Event;
-class G4Box;
+class G4VPhysicalVolume;
+class G4LogicalVolume;
 
-/// The primary generator action class with particle gun.
-///
-/// The default kinematic is a 6 MeV gamma, randomly distribued 
-/// in front of the phantom across 80% of the (X,Y) phantom size.
+/// Detector construction class to define materials and geometry.
 
-class B1PrimaryGeneratorAction : public G4VUserPrimaryGeneratorAction
+class ShieldModelDetectorConstruction : public G4VUserDetectorConstruction
 {
   public:
-    B1PrimaryGeneratorAction();    
-    virtual ~B1PrimaryGeneratorAction();
+    ShieldModelDetectorConstruction();
+    virtual ~ShieldModelDetectorConstruction();
 
-    // method from the base class
-    virtual void GeneratePrimaries(G4Event*);         
-  
-    // method to access particle gun
-    const G4ParticleGun* GetParticleGun() const { return fParticleGun; }
-    const G4GeneralParticleSource* GetParticleSource() const { return fParticleSource; }
+    virtual G4VPhysicalVolume* Construct();
+    
+    G4LogicalVolume* GetScoringVolume() const { return fScoringVolume; }
 
-  private:
-    G4ParticleGun*  fParticleGun; // pointer a to G4 gun class
-    G4GeneralParticleSource* fParticleSource;
-    G4Box* fEnvelopeBox;
+  protected:
+    G4LogicalVolume*  fScoringVolume;
 };
 
 //....oooOO0OOooo........oooOO0OOooo........oooOO0OOooo........oooOO0OOooo......
 
 #endif
+
