+++ include/ShieldModelPrimaryGeneratorAction.hh
@@ -24,40 +24,46 @@
 // ********************************************************************
 //
 //
-/// \file B1RunAction.hh
-/// \brief Definition of the B1RunAction class
+/// \file ShieldModelPrimaryGeneratorAction.hh
+/// \brief Definition of the ShieldModelPrimaryGeneratorAction class
 
-#ifndef B1RunAction_h
-#define B1RunAction_h 1
+#ifndef ShieldModelPrimaryGeneratorAction_h
+#define ShieldModelPrimaryGeneratorAction_h 1
 
-#include "G4UserRunAction.hh"
-#include "G4Accumulable.hh"
+#include "G4VUserPrimaryGeneratorAction.hh"
+#include "G4ParticleGun.hh"
+#include "G4GeneralParticleSource.hh"
 #include "globals.hh"
 
-class G4Run;
+class G4ParticleGun;
+class G4GeneralParticleSource;
+class G4Event;
+class G4Box;
 
-/// Run action class
+/// The primary generator action class with particle gun.
 ///
-/// In EndOfRunAction(), it calculates the dose in the selected volume 
-/// from the energy deposit accumulated via stepping and event actions.
-/// The computed dose is then printed on the screen.
+/// The default kinematic is a 6 MeV gamma, randomly distribued 
+/// in front of the phantom across 80% of the (X,Y) phantom size.
 
-class B1RunAction : public G4UserRunAction
+class ShieldModelPrimaryGeneratorAction : public G4VUserPrimaryGeneratorAction
 {
   public:
-    B1RunAction();
-    virtual ~B1RunAction();
+    ShieldModelPrimaryGeneratorAction();    
+    virtual ~ShieldModelPrimaryGeneratorAction();
 
-    // virtual G4Run* GenerateRun();
-    virtual void BeginOfRunAction(const G4Run*);
-    virtual void   EndOfRunAction(const G4Run*);
-
-    void AddEdep (G4double edep); 
+    // method from the base class
+    virtual void GeneratePrimaries(G4Event*);         
+  
+    // method to access particle gun
+    const G4ParticleGun* GetParticleGun() const { return fParticleGun; }
+    const G4GeneralParticleSource* GetParticleSource() const { return fParticleSource; }
 
   private:
-    G4Accumulable<G4double> fEdep;
-    G4Accumulable<G4double> fEdep2;
+    G4ParticleGun*  fParticleGun; // pointer a to G4 gun class
+    G4GeneralParticleSource* fParticleSource;
+    G4Box* fEnvelopeBox;
 };
 
-#endif
+//....oooOO0OOooo........oooOO0OOooo........oooOO0OOooo........oooOO0OOooo......
 
+#endif
