+++ init_vis.mac
@@ -1,4 +1,4 @@
-# Macro file for the initialization of example B1
+# Macro file for the initialization of example ShieldModel
 # in interactive session
 #
 # Set some default verbose
