+++ src/ShieldModelActionInitialization.cc
@@ -24,72 +24,47 @@
 // ********************************************************************
 //
 //
-/// \file B1EventAction.cc
-/// \brief Implementation of the B1EventAction class
+/// \file ShieldModelActionInitialization.cc
+/// \brief Implementation of the ShieldModelActionInitialization class
 
-#include "B1EventAction.hh"
-#include "B1RunAction.hh"
-
-#include "G4Event.hh"
-#include "G4RunManager.hh"
-#include "G4UnitsTable.hh"
-
-#include "B1Analysis.hh" 
+#include "ShieldModelActionInitialization.hh"
+#include "ShieldModelPrimaryGeneratorAction.hh"
+#include "ShieldModelRunAction.hh"
+#include "ShieldModelEventAction.hh"
+#include "ShieldModelSteppingAction.hh"
 
 //....oooOO0OOooo........oooOO0OOooo........oooOO0OOooo........oooOO0OOooo......
 
-B1EventAction::B1EventAction(B1RunAction* runAction)
-: G4UserEventAction(),
-  fRunAction(runAction),
-  fEdep(0.),
-  fLtvl(0.)
-{} 
-
-//....oooOO0OOooo........oooOO0OOooo........oooOO0OOooo........oooOO0OOooo......
-
-B1EventAction::~B1EventAction()
+ShieldModelActionInitialization::ShieldModelActionInitialization()
+ : G4VUserActionInitialization()
 {}
 
 //....oooOO0OOooo........oooOO0OOooo........oooOO0OOooo........oooOO0OOooo......
 
-void B1EventAction::BeginOfEventAction(const G4Event*)
-{    
-  fEdep = 0.;
-  fLtvl = 0.;
+ShieldModelActionInitialization::~ShieldModelActionInitialization()
+{}
+
+//....oooOO0OOooo........oooOO0OOooo........oooOO0OOooo........oooOO0OOooo......
+
+void ShieldModelActionInitialization::BuildForMaster() const
+{
+  ShieldModelRunAction* runAction = new ShieldModelRunAction;
+  SetUserAction(runAction);
 }
 
 //....oooOO0OOooo........oooOO0OOooo........oooOO0OOooo........oooOO0OOooo......
 
-void B1EventAction::EndOfEventAction(const G4Event* event)
-{ 
-  // get analysis manager
-  auto analysisManager = G4AnalysisManager::Instance();
-  
-  // fill ntuple
-  if (fEdep>0.0){
-    analysisManager->FillNtupleDColumn(0, fEdep);
-    analysisManager->FillNtupleDColumn(1, fLtvl);
-    analysisManager->AddNtupleRow();  
-  }
-  
-  // Print per event (modulo n)
-  //
-  auto eventID = event->GetEventID();
-  auto printModulo = G4RunManager::GetRunManager()->GetPrintProgress();
-  if ( ( printModulo > 0 ) && ( eventID % printModulo == 0 ) ) {
-  //if ( 1 ) {
-    G4cout << "---> End of event: " << eventID << G4endl;     
+void ShieldModelActionInitialization::Build() const
+{
+  SetUserAction(new ShieldModelPrimaryGeneratorAction);
 
-    G4cout
-       << "   Absorber: total energy: " << std::setw(7)
-                                        << G4BestUnit(fEdep,"Energy")
-       << "       total track length: " << std::setw(7)
-                                        << G4BestUnit(fLtvl,"Length")
-       << G4endl;
-  }
-
-  // accumulate statistics in run action
-  fRunAction->AddEdep(fEdep);
-}
+  ShieldModelRunAction* runAction = new ShieldModelRunAction;
+  SetUserAction(runAction);
+  
+  ShieldModelEventAction* eventAction = new ShieldModelEventAction(runAction);
+  SetUserAction(eventAction);
+  
+  SetUserAction(new ShieldModelSteppingAction(eventAction));
+}  
 
 //....oooOO0OOooo........oooOO0OOooo........oooOO0OOooo........oooOO0OOooo......
